Fix note mismatch bug
Problem: After editing a note, the fields of the note displayed in the edit screen seem to be different from the corresponding list tile note in the home screen. Once we hot reload, the changes seem to reflect

Cause: In the edit screen, we were making a shallow copy of the note that was passed from the home screen. This shallow copy still maintains a reference to the original note

When we make changes to any of the fields of the shallow copy, the fields of the original copy are affected but the home screen does not reflect this change because the home screen is not rebuilt